server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
websockets==15.0.1
sqlalchemy==2.0.41
aiosqlite==0.21.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.9.1
aiofiles==24.1.0
pillow==11.2.1
cryptography==45.0.3